Free-flight mode script enables scene navigation.
You can use WASD or the arrow keys to move forward/backwards, to the
left/right. You can also configure all these keys. Click and drag
the mouse to rotate the views.

The fallback mode is enabled according to the capabilities of the VR
device.

In desktop environments with VR, movement and rotation are available
while not presenting.

In mobile environments, movement and rotation are enabled or disabled
according to the capabilities of the device.